refactor implementations of StyleStrategy
