+++ packages/web-components/fast-element/src/polyfills.ts
@@ -1,3 +1,4 @@
+import { StyleElementStrategy } from "./components/element-controller.js";
 import type {
     FASTGlobal,
     StyleStrategy,
@@ -74,47 +75,6 @@ if (FAST.getById === void 0) {
 const supportsAdoptedStyleSheets =
     Array.isArray((document as any).adoptedStyleSheets) &&
     "replace" in CSSStyleSheet.prototype;
-
-function usableStyleTarget(target: StyleTarget): StyleTarget {
-    return target === document ? document.body : target;
-}
-
-let id = 0;
-const nextStyleId = (): string => `fast-${++id}`;
-
-export class StyleElementStrategy implements StyleStrategy {
-    private readonly styleClass: string;
-
-    public constructor(private readonly styles: string[]) {
-        this.styleClass = nextStyleId();
-    }
-
-    public addStylesTo(target: StyleTarget): void {
-        target = usableStyleTarget(target);
-
-        const styles = this.styles;
-        const styleClass = this.styleClass;
-
-        for (let i = 0; i < styles.length; i++) {
-            const element = document.createElement("style");
-            element.innerHTML = styles[i];
-            element.className = styleClass;
-            target.append(element);
-        }
-    }
-
-    public removeStylesFrom(target: StyleTarget): void {
-        const styles: NodeListOf<HTMLStyleElement> = target.querySelectorAll(
-            `.${this.styleClass}`
-        );
-
-        target = usableStyleTarget(target);
-
-        for (let i = 0, ii = styles.length; i < ii; ++i) {
-            target.removeChild(styles[i]);
-        }
-    }
-}
 
 if (!supportsAdoptedStyleSheets) {
     FAST.getById(/* KernelServiceId.styleSheetStrategy */ 5, () => StyleElementStrategy);
